+++ tests/testthat/helper_nonNA_cells.R
@@ -1,0 +1,3 @@
+nonNA_cells <- function(x) {
+  terra::cells(is.na(x), 0)[[1]]
+}